refactor(mambavesselnet): simplify — consolidate MockMamba, add _build_config extras
- tests/v2/unit/conftest.py: extract shared MockMamba (was duplicated in 2 test files)
- test_mambavesselnet_blocks.py + test_mambavesselnet_adapter.py: import from conftest
- TestMambaVesselNetAdapterForward: extract _make_adapter() helper (matches Interface class)
- MambaVesselNetAdapter.get_config(): expose feature_dims, d_state, d_conv, expand in extras
- Store _feature_dims/_d_state/_d_conv/_expand as instance vars in __init__
